+++ wfc3tools/wf3rej.py
@@ -1,22 +1,11 @@
-from __future__ import print_function
-
 # STDLIB
 import os.path
 import subprocess
 
-# get the auto update version for the call to teal help
-from .version import __version_date__, __version__
 
 # STSCI
 from stsci.tools import parseinput
 from .util import error_code
-
-try:
-    from stsci.tools import teal
-    has_teal = True
-except ImportError:
-    has_teal = False
-    print("Teal not available")
 
 __taskname__ = "wf3rej"
 
@@ -119,55 +108,3 @@ def wf3rej(input, output="", crrejtab="", scalense="", initgues="",
             print("Unknown return code found!")
             ec = return_code
         raise RuntimeError("wf3rej.e exited with code {}".format(ec))
-
-
-def help(file=None):
-    helpstr = getHelpAsString(docstring=True)
-    if file is None:
-        print(helpstr)
-    else:
-        if os.path.exists(file):
-            os.remove(file)
-        f = open(file, mode='w')
-        f.write(helpstr)
-        f.close()
-
-
-def getHelpAsString(docstring=False):
-    """Return documentation on the 'wf3ir' function. Required by TEAL."""
-
-    install_dir = os.path.dirname(__file__)
-    htmlfile = os.path.join(install_dir, 'htmlhelp', __taskname__ + '.html')
-    helpfile = os.path.join(install_dir, __taskname__ + '.help')
-    if docstring or (not docstring and not os.path.exists(htmlfile)):
-        helpString = ' '.join([__taskname__, 'Version', __version__,
-                               ' updated on ', __version_date__]) + '\n\n'
-        if os.path.exists(helpfile) and has_teal:
-            helpString += teal.getHelpFileAsString(__taskname__, __file__)
-    else:
-        helpString = 'file://' + htmlfile
-
-    return helpString
-
-
-def run(configobj=None):
-    """TEAL interface for the ``wf3rej`` function."""
-
-    wf3rej(configobj['input'],
-           output=configobj['output'],
-           crrejtab=configobj['crrejtab'],
-           scalense=configobj['scalense'],
-           initgues=configobj['initgues'],
-           skysub=configobj['skysub'],
-           crsigmas=configobj['crsigmas'],
-           crradius=configobj['crradius'],
-           crthresh=configobj['crthresh'],
-           badinpdq=configobj['badinpdq'],
-           crmask=configobj['crmask'],
-           shadcorr=configobj['shadcorr'],
-           verbose=configobj['verbose'],)
-
-
-# This replaces the help for the function which is also printed in the HTML
-# and TEAL
-wf3rej.__doc__ = getHelpAsString(docstring=True)